requirements.txt
Flask==2.0.3
google-api-python-client==2.31.0
openai==0.10.2
python-dotenv==0.19.2